+++ packages/client/wallets/smart-wallet/src/evm/rpc.ts
@@ -16,7 +16,7 @@ export const ALCHEMY_RPC_SUBDOMAIN: Partial<Record<SmartWalletChain, string>> = 
     "arbitrum-sepolia": "arb-sepolia",
 };
 
-function getAlchemyRPC(chain: SmartWalletChain): string {
+export function getAlchemyRPC(chain: SmartWalletChain): string {
     return `https://${ALCHEMY_RPC_SUBDOMAIN[chain]}.g.alchemy.com/v2/${ALCHEMY_API_KEY}`;
 }
